[RED] get all homes from home services

+++ SmartHome/Domain/Home.cs
@@ -10,8 +10,8 @@ public class Home
     public int Id { get; set; }
     public string Street { get; set; }
     public int DoorNumber { get; set; }
-    public int Latitude { get; set; }
-    public int Longitude { get; set; }
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
     
     public List<Member> Members
     {
@@ -21,7 +21,7 @@ public class Home
     
     public List<Device> Devices { get; set; }
     
-    public Home(string street, int doorNumber,int latitude,int longitude)
+    public Home(string street, int doorNumber,double latitude,double longitude)
     {
         Street = street;
         DoorNumber = doorNumber;
